backend/flask/requirements.txt
flask==2.2.5
python-dotenv==1.0.0
pymongo==4.4.1
requests
flask-cors
pyopenssl